requirements.txt
python==3.9.6
request==2.26.0
